third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-bias-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'bias' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('bias') is "0.5"
PASS convolveMatrixElement.getAttribute('bias') is "0"
PASS successfullyParsed is true

TEST COMPLETE
